+++ dst/srcs/stackmin.c
@@ -6,18 +6,19 @@
 /*   By: tkano <tkano@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/04 19:08:58 by tkano             #+#    #+#             */
-/*   Updated: 2021/10/04 19:44:38 by tkano            ###   ########.fr       */
+/*   Updated: 2021/10/06 21:14:23 by tkano            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "pushswap.h"
 
-long	stackmin(s_stack **a, int *loc, long len)
+long	stackmin(s_stack **a, int *loc, int len)
 {
 	long	min;
 	int		i;
 
 	i = 1;
+	//printf("check3\n");
 	while ((*a)->prev->value != DUMMY)
 	{
 		*a = (*a)->next;
@@ -34,5 +35,6 @@ long	stackmin(s_stack **a, int *loc, long len)
 		else
 			i++;
 	}
+	//printf("check4, min: %ld\n", min);
 	return (min);
 }
